Refactor: (Feedback) RecruitmentStatus 클래스 패키지 이동 (enrollment 하위로 이동)
