+++ src/main/java/nextstep/courses/domain/enrollment/PaidSessionEnrollment.java
@@ -1,7 +1,6 @@
 package nextstep.courses.domain.enrollment;
 
 import nextstep.courses.domain.enrollment.engine.SessionEnrollment;
-import nextstep.courses.domain.status.RecruitmentStatus;
 import nextstep.courses.domain.student.SessionStudent;
 import nextstep.courses.exception.SessionFeeMismatchException;
 import nextstep.payments.domain.Payment;
