+++ src/main/java/nextstep/courses/domain/enrollment/RecruitmentStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.status;
+package nextstep.courses.domain.enrollment;
 
 import nextstep.courses.exception.SessionRecruitmentStatusInvalidException;
